Django_Blog/requirements.txt
altgraph==0.17.4
amqp==5.3.1
asgiref==3.8.1
beautifulsoup4==4.13.3
billiard==4.2.1
bs4==0.0.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.1
Django==5.1.7
django-filter==25.1
django-session-security==2.6.7
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
djangorestframework_simplejwt==5.5.0
idna==3.10
kombu==5.4.2
mysql-connector==2.2.9
mysql-connector-python==9.2.0
mysqlclient==2.2.7
oauthlib==3.2.2
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
prompt_toolkit==3.0.50
pycparser==2.22
pycryptodomex==3.21.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyJWT==2.9.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
pyzipper==0.3.6
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.8.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
